requirements.txt
appnope==0.1.4
asttokens==3.0.1
beautifulsoup4==4.14.3
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
comm==0.2.3
contourpy==1.3.3
curl_cffi==0.13.0
cycler==0.12.1
debugpy==1.8.20
decorator==5.2.1
executing==2.2.1
fonttools==4.62.1
frozendict==2.4.7
idna==3.11
ipykernel==7.2.0
ipython==9.10.1
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jupyter_client==8.8.0
jupyter_core==5.9.1
kiwisolver==1.5.0
matplotlib==3.10.8
matplotlib-inline==0.2.1
multitasking==0.0.12
nest-asyncio==1.6.0
numpy==2.4.3
packaging==26.0
pandas==3.0.1
parso==0.8.6
peewee==4.0.3
pexpect==4.9.0
pillow==12.1.1
platformdirs==4.9.4
prompt_toolkit==3.0.52
protobuf==7.34.1
psutil==7.2.2
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==3.0
Pygments==2.19.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
pytz==2026.1.post1
PyYAML==6.0.3
pyzmq==27.1.0
requests==2.33.0
six==1.17.0
soupsieve==2.8.3
stack-data==0.6.3
tornado==6.5.5
traitlets==5.14.3
typing_extensions==4.15.0
urllib3==2.6.3
wcwidth==0.6.0
websockets==16.0
yfinance==1.2.0
